backend/requirements.txt
Flask
boto3
python-dotenv
pytest
Flask-Testing
moto
flasgger
